requirements.txt
alembic==1.13.1
blinker==1.7.0
Bootstrap-Flask==2.3.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.5.0
elastic-transport==8.11.0
elasticsearch==8.11.1
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Login @ git+https://github.com/maxcountryman/flask-login.git@ea80968c10f3d521ea4983652f183aa42e2f44fc
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Moment==1.0.5
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.0
MarkupSafe==2.1.3
mysqlclient==2.2.1
packaging==23.2
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
urllib3==2.1.0
vercel==0.2.1
Werkzeug==3.0.1
WTForms==3.1.2
